Export Attachments
Summary
There are no geoprocessing tools that allow users to export and save all attachments locally. The instructions provided below describe a possible solution.
The following script iterates through the entire attachment table of a single feature class and copies all of the attachments (saved as BLOBs, or binary large objects) to file.